CHORE: Add haproxy-internal container to persist stick-table data
Public HAProxy instance is being infrequently restarted to update SSL certs or
Tor exit relays lists. By adding an internal only HAProxy instance we can keep
stick-tables in sync and hence persist them.
Next to it, change the stick-tables type from `ip` to `ipv6`.

+++ contrib/docker-compose/standalone/haproxy-entrypoint.sh
@@ -1,9 +1,16 @@
 #!/bin/sh
+export HOSTNAME=$(hostname)
+export HOST_IP=$(hostname -i)
 
-haproxy "$@" &
+RELOAD_INTERVAL=${RELOAD_INTERVAL:-1800}
 
-trap exit TERM
-while :; do
-  sleep 1800
-  ps | awk '{ if($4 == "haproxy") {print $1} } ' | xargs kill -HUP
-done
+if [ "x${RELOAD_INTERVAL}" != "x0" ]; then
+  haproxy "$@" &
+  trap exit TERM
+  while true; do
+    sleep $RELOAD_INTERVAL
+    ps | awk '{ if($4 == "haproxy") {print $1} } ' | xargs kill -HUP
+  done
+else
+  exec haproxy "$@"
+fi
